% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BakerHughesRandD.R
\docType{data}
\name{BakerHughesRandD}
\alias{BakerHughesRandD}
\title{Baker Hughes Corporation data on R&D}
\format{
An object of class \code{character} of length 75.
}
\usage{
data(BakerHughesRandD)
}
\description{
Data on hydraulic fracturing R&D projects.
}
\examples{
library(MultiplierDEA)

resBH<-MultiplierDEA::DeaMultiplierModel(BakerHughesRandD$XBH,
                          BakerHughesRandD$YBH,
                          rts = "vrs",
                          orientation="input")
}
\references{
Thore, “Technology Commercialization,” Springer, 2002, pp. 424.
}
\keyword{datasets}
