% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_helper.R
\name{DrawIOdiagram}
\alias{DrawIOdiagram}
\title{Draw DEA Input-Output Diagram}
\usage{
DrawIOdiagram(Xnames = "1", Ynames = "1", Modeldescrip = "DEA")
}
\arguments{
\item{Xnames}{Names of Inputs}

\item{Ynames}{Names of Outputs}

\item{Modeldescrip}{Model Description}
}
\value{
Input-Output Diagram as PNG
}
\description{
Draw DEA Input-Output Diagram
}
\examples{
DrawIOdiagram (c("X1 (FTE)" ), c("Y1 ($M)"), '"\nBCC-IO\n "')
}
