% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malm_graphical.R
\docType{data}
\name{malm_graphical}
\alias{malm_graphical}
\title{Malmquist Productivity Index Data}
\format{
An object of class \code{data.frame} with 6 rows and 4 columns.
}
\usage{
data(malm_graphical)
}
\description{
2D data for demonstrating Malmquist Productivity Index analysis
}
\examples{
malm_graphical

library(deaR)
#Process data into the format required by deaR
malm_gr_data_pro <-
  read_malmquist(malm_graphical,
                 percol = 1,   # Use Col1 for period (year)
                 dmus = 2,     # Col2 has DMU identifier
                 inputs = 3,   # Col3 has the input
                 outputs = 4,  # Col4 has the output
                 arrangement = "vertical")

result_gr <- malmquist_index(malm_gr_data_pro,
                             orientation = "oo",
                             rts = "crs",
                             type1 = "cont",
                             type2 = "fgnz")

mi_gr <- result_gr$mi

res_mi_gr <- t(rbind(result_gr$mi, result_gr$ec,
                     result_gr$tc ))
colnames(res_mi_gr) <- c("MPI", "PEch",  "TC")
res_mi_gr
}
\keyword{datasets}
