% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_helper.R
\name{poscol}
\alias{poscol}
\title{Filters Out Zero Columns}
\usage{
poscol(df, cutoff = 0)
}
\arguments{
\item{df}{matrix}

\item{cutoff}{Cutoff value for passing columns}
}
\value{
df matrix w/o all zero columns
}
\description{
Filters Out Zero Columns
}
\examples{
poscol(matrix(c(1,2,0,0,5,0),ncol=3)) # Filters 0, 0 column
}
