% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_lic_2007.R
\docType{data}
\name{univ_lic_2007}
\alias{univ_lic_2007}
\title{University Licensing Income Data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(univ_lic_2007)
}
\description{
Data from the Association of University Technology Managers
of 50 universities on research spending along with outcomes
such licensing income
increments over eight hours.
}
\examples{
colnames(univ_lic_2007) <- c("University",
                             "LicInc",
                             "LicOpt",
                             "Startup",
                             "PatFile",
                             "PatIss",
                             "Spend")
head(univ_lic_2007[,1:4])
}
\references{
Anderson, Daim, and Lavoie, “Measuring the Efficiency of University Technology Transfer,” Technovation, (2007), pp. 306-18.

(\href{https://www.sciencedirect.com/science/article/abs/pii/S0166497206001076}{Elsevier})
}
\keyword{datasets}
