% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_helper.R
\name{DEAnames}
\alias{DEAnames}
\title{Generates Generic Names for DEA}
\usage{
DEAnames(NX, NY, ND)
}
\arguments{
\item{NX}{Number of inputs}

\item{NY}{Number of outputs}

\item{ND}{Number of decision making units}
}
\value{

}
\description{
Generates Generic Names for DEA
}
\examples{
DEAnames (1,2,3) # 1-input, 2-output, 3 DMUs
}
